pyproject.toml
[tool.poetry]
name = "spotify-scraper"
version = "0.1.0"
description = ""
authors = ["NINTAI DICK <nintaidick67@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
playwright = "^1.45.1"
django = "^5.1"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
